requirements.txt
requests
colorama
websockets